+++ assets/js/site.js
@@ -7,7 +7,7 @@ import { FloatingPanel } from "corex/floating-panel"
 import { Toast } from "corex/toast"
 import { Select } from "corex/select"
 import { Toggle } from "corex/toggle"
-import { initLenis } from "./lenis.js"
+import { initLenis } from "./init-lenis.js"
 import { initLanding } from "./landing.js"
 import { initWaitlistForm } from "./waitlist.js"